@@ -1,63 +1,103 @@
-import axios, {AxiosError, AxiosResponse} from 'axios';
-import {
-  requestDefaultInterceptors,
-  responseDefaultInterceptors,
-} from './interceptors';
-import {accessToken} from "@/utils/storage/token";
-import {getLocaleFromUrl} from "@/i18n/routing";
+import axios from 'axios'
+import type { AxiosError, AxiosResponse, InternalAxiosRequestConfig } from 'axios'
+import { CryptoUtils } from '@/utils/crypto'
+import { accessToken } from '@/utils/storage/token'
+import { getLocale } from '@/i18n'
 
-// 初始化实例
+// 创建 axios 实例
 const service = axios.create({
   timeout: 300000,
   headers: {
-    'Accept': 'application/json',
-    'X-Requested-With': 'XMLHttpRequest',
-  },
-});
+    Accept: 'application/json',
+    'X-Requested-With': 'XMLHttpRequest'
+  }
+})
 
 // 请求拦截器
-service.interceptors.request.use(requestDefaultInterceptors);
-
-// 请求拦截器 必须放最后
 service.interceptors.request.use(
-  config => {
-    return config;
+  async (config: InternalAxiosRequestConfig) => {
+    const body = config.headers['Content-Type'] === 'multipart/form-data' ? {} : config.data
+
+    const url = window.location.pathname
+    config.headers['Accept-Language'] = getLocale(url)
+
+    // 签名
+    const {
+      timestamp: _timestamp,
+      nonce: _nonce,
+      signature: _signature
+    } = await CryptoUtils.createSignedRequest({
+      method: (config.method || '').toUpperCase(),
+      path: config.url || '',
+      privateKey: import.meta.env.VITE_API_SIGN_IN_PRIVATE_KEY || '',
+      body: body || {}
+    })
+
+    config.headers.Authorization = `${accessToken.getTokenType()} ${accessToken.getToken()}`
+    config.headers.Nonce = _nonce
+    config.headers.Sign = _signature
+    config.headers.Timestamp = _timestamp
+
+    return config
   },
   (error: AxiosError) => {
-    console.info('axios request error: ', error);
-    return Promise.reject(error);
-  },
-);
+    console.error('axios request error:', error)
+    return Promise.reject(error)
+  }
+)
 
-// 响应拦截器 必须放第一个
-// 统一抛出错误，方便第二个中间件捕获错误后 统一进行提示
+// 响应拦截器 - 处理认证错误
 service.interceptors.response.use(
   (response: AxiosResponse) => {
     if (response.status === 200) {
-      return response;
+      return response
     }
     if (response?.status === 401 || response?.status === 403) {
-      accessToken.remove();
-
-      // return Promise.reject({
-      //   ...response,
-      //   isAuthError: true,
-      // });
-
-      if (typeof window !== "undefined") {
-        const pathname = window.location.pathname;
-        const locale = getLocaleFromUrl(pathname);
-        window.location.href = `${locale}/auth/login?redirect=${pathname}`;
-      }
+      handleAuthError()
     }
-    return Promise.reject(response);
-  },
-);
+    return Promise.reject(response)
+  }
+)
 
-// 响应拦截器
+// 响应拦截器 - 提取数据
 service.interceptors.response.use(
-  responseDefaultInterceptors.success,
-  responseDefaultInterceptors.error,
-);
+  (response: AxiosResponse) => {
+    return response.data
+  },
+  (error: AxiosError) => {
+    // 超时处理
+    if (!error.response && error?.message.indexOf('timeout') !== -1) {
+      return Promise.reject(error)
+    }
 
-export default service;
+    // HTTP 状态码错误处理
+    switch (error?.response?.status) {
+      case 401:
+      case 403:
+        handleAuthError()
+        break
+      case 404:
+        console.error('请求地址不存在')
+        break
+      case 501:
+      case 502:
+      case 503:
+      case 504:
+        console.error('服务端错误')
+        break
+    }
+
+    return Promise.reject(error)
+  }
+)
+
+// 处理认证错误
+function handleAuthError() {
+  accessToken.remove()
+  const url = window.location.pathname
+  if (typeof window !== 'undefined') {
+    window.location.href = `/auth/login?redirect=${url}`
+  }
+}
+
+export default service
